Using hosts pattern for inventory var resolution :sparkles:
Prior behavior was to resolve variables by hostname alone.

This meant that if a var was declared in multiple groups
for a given host you would be unable to know which group var
would be used.

This commit uses the user provided pattern to ensure
groups in the pattern have priority.